Update Dockefile to make image build faster
This commit seperate
COPY . .
command to

COPY package.json yarn.lock .
COPY . .

It will allow us to use cache to rebuild the image.
Now, if we change files of this WORKDIR, although the yarn dependencies
will be reinstalled, but it won't affect the
COPY . . command, becuase COPY . . is the downstream layer right now.

Reference:
    - [Docker getting started - Part 9 image-building best practices](https://docs.docker.com/get-started/09_image_best/)
    - [Dockerizing a Node.js web app](https://nodejs.org/en/docs/guides/nodejs-docker-webapp/)